ERROR in bgp_router_refs for bgp-as-a-service in config
UI payload used to contain the old bgp_router_refs which was get on last
bgp-as-aservice object, as while sending the data to nodejs, it does not do
again GET on config object, so, there is inconsistent behavior when it comes
to nodejs. Nodejs sees a diff on bgp_router_refs between UI payload and
whatever is available in Config. So it sends a update request on bgp-as-as-service
object with wrong bgp_router_refs info which leads the issue.

Fix: As nodejs does not need to update this object, so remove from UI payload.

closes-jira-bug: CEM-19989
Change-Id: I1c3f8fed81bcd523f2ff443fcd33efeb1cec62c5