Fix for DNS untick option in Virtual Network Subnet section not working in TF UI
Problem : User created a network with disabled DNS, on edit form of network the DNS is showing as enabled.

Fix : The DNS checkbox will be disabled/enabled as per the user selection. This fix done for ipv4 and ipv6 addresses.
Also we verified the default gateway value is getting as 0.0.0.0 for ipv4 and  :: for ipv6 subnets( without breaking the functinality on CEM-15847).

closes-jira-bug: CEM-24454

Signed-off-by: jismariag <jismariag@juniper.net>
Change-Id: I1840445000c2ec35b0090bad37c17a385c6e9eae